// src/enums/connection-type.d.ts
export declare enum ConnectionType {
    POWER_INPUT = 0,
    POWER_OUTPUT = 1,
    GAS_INPUT = 2,
    GAS_OUTPUT = 3,
    LIQUID_INPUT = 4,
    LIQUID_OUTPUT = 5,
    LOGIC_INPUT = 6,
    LOGIC_OUTPUT = 7,
    SOLID_INPUT = 8,
    SOLID_OUTPUT = 9,
    NONE = 10
}
//# sourceMappingURL=connection-type.d.ts.map